+++ src/main/java/com/diecast/diecast_back/repository/LinhaMiniaturaRepository.java
@@ -4,7 +4,7 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.diecast.diecast_back.model.LinhaMiniatura;
 
-public interface LinhaMiniaturaRepository extends JpaRepository<LinhaMiniatura, Short> {
+public interface LinhaMiniaturaRepository extends JpaRepository<LinhaMiniatura, Long> {
 
 	Optional<LinhaMiniatura> findByNome(String nome);
